+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     none is available.
     - FZF, Unite and Denite support multiple choices and selection preview.
 - `@task.options_multi` - like `task.options` but allows multiple choices.
+- Generator style for `@task.options`.
 
 ### Changed
 - Look the directory tree for a tasks file.